clean and fix sdk
