--- sdk/circuits/dsc_4096_vkey.json
@@ -1,99 +1,0 @@
-{
- "protocol": "groth16",
- "curve": "bn128",
- "nPublic": 2,
- "vk_alpha_1": [
-  "20491192805390485299153009773594534940189261866228447918068658471970481763042",
-  "9383485363053290200918347156157836566562967994039712273449902621266178545958",
-  "1"
- ],
- "vk_beta_2": [
-  [
-   "6375614351688725206403948262868962793625744043794305715222011528459656738731",
-   "4252822878758300859123897981450591353533073413197771768651442665752259397132"
-  ],
-  [
-   "10505242626370262277552901082094356697409835680220590971873171140371331206856",
-   "21847035105528745403288232691147584728191162732299865338377159692350059136679"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_gamma_2": [
-  [
-   "10857046999023057135944570762232829481370756359578518086990519993285655852781",
-   "11559732032986387107991004021392285783925812861821192530917403151452391805634"
-  ],
-  [
-   "8495653923123431417604973247489272438418190587263600148770280649306958101930",
-   "4082367875863433681332203403145435568316851327593401208105741076214120093531"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_delta_2": [
-  [
-   "5689878758968408864570330769586116386424915777760869435924001045372426387042",
-   "14835593574773248144471339475507473753186624066904877663229402603601341881110"
-  ],
-  [
-   "2051149603231436076479384106975809917151570357124249160820058911298166997663",
-   "628452860859295088428620228489450512026200467380394004198146223506590955249"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_alphabeta_12": [
-  [
-   [
-    "2029413683389138792403550203267699914886160938906632433982220835551125967885",
-    "21072700047562757817161031222997517981543347628379360635925549008442030252106"
-   ],
-   [
-    "5940354580057074848093997050200682056184807770593307860589430076672439820312",
-    "12156638873931618554171829126792193045421052652279363021382169897324752428276"
-   ],
-   [
-    "7898200236362823042373859371574133993780991612861777490112507062703164551277",
-    "7074218545237549455313236346927434013100842096812539264420499035217050630853"
-   ]
-  ],
-  [
-   [
-    "7077479683546002997211712695946002074877511277312570035766170199895071832130",
-    "10093483419865920389913245021038182291233451549023025229112148274109565435465"
-   ],
-   [
-    "4595479056700221319381530156280926371456704509942304414423590385166031118820",
-    "19831328484489333784475432780421641293929726139240675179672856274388269393268"
-   ],
-   [
-    "11934129596455521040620786944827826205713621633706285934057045369193958244500",
-    "8037395052364110730298837004334506829870972346962140206007064471173334027475"
-   ]
-  ]
- ],
- "IC": [
-  [
-   "14532585603301627491293557600415273871129994972736322653869196794622673732921",
-   "18882397311658692526609112965864523608034856647654838257030831022893270548325",
-   "1"
-  ],
-  [
-   "4186298633367159062520988354552715829159959996352603762446141935317873102425",
-   "3924212783150620566854971503130830169563464952561919541343311327675656189646",
-   "1"
-  ],
-  [
-   "4437362653636713552232717650847448346211317604147264938993567382296904710753",
-   "2938386579388053271692442071404949636853984120968379123259292205410760857324",
-   "1"
-  ]
- ]
-}
