--- sdk/index.ts
@@ -1,21 +1,0 @@
-import { OpenPassportWeb2Verifier } from './OpenPassportWeb2Verifier';
-import { OpenPassportWeb2Inputs } from './OpenPassportWeb2Verifier';
-import { OpenPassportVerifierReport } from './OpenPassportVerifierReport';
-// import { OpenPassportWeb3Verifier } from './OpenPassportWeb3Verifier';
-import { countryCodes } from '../common/src/constants/constants';
-import { QRCodeGenerator } from './QRCodeGenerator';
-import { AppType } from '../common/src/utils/appType';
-import { OpenPassportProverVerifier, OpenPassportProverInputs } from './OpenPassportProverVerifier';
-
-export {
-    OpenPassportWeb2Verifier,
-    // OpenPassportWeb3Verifier,
-
-    OpenPassportWeb2Inputs,
-    OpenPassportVerifierReport,
-    QRCodeGenerator,
-    AppType,
-    countryCodes,
-    OpenPassportProverVerifier,
-    OpenPassportProverInputs
-}
